fixing up chat and world references for 1.6
also entityLivingBase and some fluid stuff

+++ common/logisticspipes/interfaces/routing/IRequestLiquid.java
@@ -7,5 +7,5 @@ import logisticspipes.utils.LiquidIdentifier;
 public interface IRequestLiquid {
 	IRouter getRouter();
 	void sendFailed(LiquidIdentifier value1, Integer value2);
-	float canSink(FluidStack f); //returns the amount actually sunk (ie, free space for the liquid)
+	int canSink(FluidStack f); //returns the amount actually sunk (ie, free space for the liquid)
 }
